Architecture and API rework
Previously qtopcua
- was trying to do too much. It was a C++ API and a QML API (albeit not a
nice one). This patch separates this clearly -- there is a C++ API and
there will be a true declarative API in the future.
- did not implement the d-pointer pattern, making binary compatibility
impossible.
- did not separate between a client API (intended for an application
developer) and a backend API. This made a lot of internals show up in
the public API which must be hidden from the "public". This also meant
a lot of virtuals in the public API which is also not how a Qt-API
should be structured.
- implemented a way where the plugin provider would somehow choose the
"best" backend plugin. This is not what the application developer
typically wants. Now the application has more control and also needs
to explicitly select a backend plugin.
- placed a lot of "node" functionality in the QtOpcUaClient (read,
write, ...) making the QtOpcUaNode akin to an empty shell. This patch
moves this node functionality into QtOpcUa node where it is more natural.
This also prepares for the introduction of async method calls. The
intention there is to have the node object as the receiver of async
completion calls on that node.
- was trying to manage various dynamically allocated objects for the
client (e.g. nodes, subscriptions). This was due to qtopcua also
trying to be useable from the QML side. This problem was that this
was very intransparent and also lead to a complex implementation. This
patch is vastly simplifying the resource management. Dynamically
allocated objects are owned by the client who is responsible to
deallocate them.
- was unclear if subscriptions should be exposed to the application
programmer or not. This patch does expose subscriptions in the
public API. The reason is that one wants to be able to configure
a subscrption independent from the contains monitored items
(e.g. publishing intervals vs. sampling intervals).

Change-Id: Icb37e89e46a64dd9e9bd2c45abaff503e2eca8a5
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>