fix tests against latest changes
Need to adapt due to async call changes. Also now builds and runs on
Windows.
The default stack size needs to be increased on Windows, otherwise the
call to create the default nodes in freeopcua will cause a stack
overflow. This effects debug and release builds, and there is also a bug
report on github about this.

Change-Id: I0d818b2cb268c704a9ba6c7965db242163dacfda
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>