Remove QOpcuaClient::set[Error|State] and ensure their atomicity
Those functions should not be public. The API user should not be able
to influence the state or error.

The functions merely existed to facilitate internal state tracking.
The new approach does not need those public functions anymore.
Furthermore the implementation created situations where a state change
might occurr while the appropriate internal error condition was not set
or vice versa. A single function ensures that all internal state and error
changes are atomic before the various state and error signals are emitted.

Change-Id: I05fea292be227a41ce28268bade2ff9a1dab10ab
Reviewed-by: Jochen Seemann <seemann.jochen@gmail.com>
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>