Convert FreeOpcua connection mgmt to dedicated Thread usage
The thread has the advantage that the OpcUa::Client instance is in a
dedicated thread and cleans itself up when the thread disappears.
The existing node and subscription implementations still access the
OpcUa::Client instance without any consideration of thread affinity
but that is another patch and wasn't working prior to this patch anyway.

Change-Id: Id9141967e72456944dbee486814fcdd7465577ff
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>