Make the individual test cases more atomic
We cannot rely on the order of tests being executed. In particular
the connection is established and all subsequent test cases assume
the connection being alive. This defeats alternative run methods for
unit tests where the name of a specific test function is passed
as an argument and only the mentioned function is executed.

In addition, the connect & disconnect test functions were merged into
the same test case.

Last but not least QSignalSpy is used to verify the exact order or
signals and permits a more linear execution path because some slots
for signal handling are no longer necessary.

Change-Id: I5dba9653f04e3fc57cf3b7af9cb7e6dbee40a5aa
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>