Fix test failure when using freeopcua client
FAIL!  : Tst_QOpcUaClient::dataChangeSubscription(freeopcua) 'valueSpy.at(0).at(0).toDouble() == (double) 42' returned FALSE. ()
   Loc: [tst_client.cpp(379)]

It turns out that a data subscription sends an initial notification update
right after establishing the subscription. The first value is the current
value of the node. This patch ensures that this initial update is
suppressed and as a matter of fact any other update that does not change
the value is suppressed as well.

Change-Id: Iec856ab0b76952d312b39a170ba235cbc1497042
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>