Fix the handling of byte strings in QOpen62541ValueConverter
According to part 6 5.1.2 of the OPC UA specification, a byte string is
"A sequence of octets".
According to this definition, null bytes in a byte string are perfectly
legal and must be accepted as input. This means that it is not possible
to use string copying function without length parameter.

Test cases including byte strings have been extended to ensure the
proper handling of null bytes.

Change-Id: I2dd1677b47a2590e2c4a06aaa83cf9839bd754d2
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>