Add support for the GUID type
This change adds support for the GUID type to both backends.
Scalar and array GUID values can be written and read.
The correct type is returned when type() is called on a variable node
with a GUID value.

The writeArray, readArray, writeScalar and readScalar tests have been
extended to verify the correct reading and writing of GUID values.

Change-Id: I29ec98e69161e03a7df941f998e5db5667400876
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>