Unbreak and re-enable the SByte part of the readArray test
Probably due to a copy & paste error, the type of a QVariant belonging
to the value of the Byte node was checked instead of the SByte value.
The type mismatch detected by QVERIFY seems to have been misattributed
to the backend and the test was skipped for both backends.

If the right value is checked, the test passes for both backends.

Change-Id: Idc0d276ae6a8612ebf9275216633b8599fae9d7a
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>