Separate conversion and array/scalar handling in the value converters
The current version of the value converters have lots of duplicated
code for scalar and array values. Supporting alias types like
StatusCode and XmlElement would require copying the code from the
underlying types.

This change implements a template based solution to value conversion
which solves the duplicate code issue and enables easy conversion of
alias types by providing a conversion for scalar values and handling
the difference between array and scalar in a separate method.

Change-Id: I2b8be27078419d9fb9f2b6ea7e23912b5c90d871
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>
Reviewed-by: André Hartmann <aha_1980@gmx.de>