Split QFreeOpcUaClient and QFreeOpcUaWorker in separate files
Both classes were in the same implementation file and
needed to be split up in preparation for the async API.
As long as the worker class is hidden in the client source,
outside usage is not possible, because the type would be
incomplete. We require direct access to the worker if we want
to move SDK specific functionality from QFreeOpcUaNode
into the async backend.
Removed unused includes in the process.

Change-Id: Ie13c0e5cd290fb50f26e302856fe89429c7f196d
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>