Add support for the QualifiedName type
QualifiedName is mainly used for the BrowseName attribute.
It can also be used as type for the value attribute of a variable node.

This patch adds support for QualifiedName values to both backends.
QualifiedName scalar and array variable nodes are added to the test
server, the scalar and array tests are extended to verify correct value
conversion.

Change-Id: Ic678c7c774d6594fde8088450d00bfc829ea9743
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>