Add ExtensionObject handling to support types from OPC-UA part 8, 5.6
This patch adds a generic OPC UA Binary DataEncoding handler class to QtOpcUa.
QOpen62541ValueConverter uses this class to convert Range, EUInformation,
ComplexNumber, DoubleComplexNumber, AxisInformation and XV.
As Range/EUI values can now be read using the generic method, the separate
methods for Range and EUInformation reading are removed.

Variable nodes with values of the new types have been added to the test server.
The scalar and array tests have been extended to check for correct conversion.

The freeopcua library does not support ExtensionObject variants, so there is
no implementation in the value converter and the tests are skipped.

Change-Id: Iec312e499d21c8cfa25fcf6be866ee432a883ed4
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>