Add async browse support
This is a replacement for the synchronous childrenIds().
browseChildren() allow filtering for node classes and reference type.
It returns detailed information on the children.

The filtering makes it possible to find properties of a node.

The following information on every child node matching the criteria is returned:
- Reference type
- Node Id
- BrowseName
- DisplayName
- NodeClass
This information is sufficient to create a tree structure without having to
read attributes from each child node.

Change-Id: Id488ec430e393f1881de483c24d392b8cb607989
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>