Add water tank simulation example
This patch adds a server application based on open62541 which contains
two water tanks, a pump and a valve. Water can be pumped from the first
tank into the second tank. If the valve is opened, water is removed from
the second tank. Both operations have a configurable setpoint.

This patch also adds a QML based HMI application which interacts with the
simulation running on the server. The setpoint can be set using a slider,
buttons start and stop the pump, open the valve or reset the simulation.

Change-Id: I3b452dfc5dd681f2e6c78cf6f6815b35d8c560b1
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>