Fix DateTime conversion for uacpp and freeopcua
The OPC UA representation for DateTime is a uint64 containing
the milliseconds since 1601-01-01, so the conversion in uacpp
is wrong by many years.

Freeopcua and uacpp both offer DateTime::toTime_t, but a time_t
only has an accuracy of one second. This is why manual conversion
like in the open62541 backend must be used.

Change-Id: I644f061478dfab5266615ba91f45e179a8c2581a
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>