Don't create default constructed entries in m_attributeMapping (uacpp)
Using QHash::operator[] with a key that does not exist inserts a default
constructed value.
modifyMonitoring() and disableMonitoring() in uacppbackend.cpp use
m_attributeMapping[handle][attr] to check if an entry exists. Calling
these methods for handle/attribute pairs that are not being monitored
fills m_attributeMapping with default constructed values.
Checking if the QHash contains the expected values avoids that.

Change-Id: I80804a1b3af5d7f9fa86a14c6f9a371a36614f6e
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>