Fix possible crash in the open62541 backend
QOpen62541Node::enableMonitoring() passes the node id member of the
node to the backend. If a string node id is used and the node is
deleted while or before enableMonitoring in the backend is running,
a crash occurs.
Passing a copy of the node id to the backend fixes the issue.

Change-Id: I0fca4150226c24182022afb95a72dab98f3d712a
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>