Disable the freeopcua tests when running in the CI
The freeopcua tests randomly run into timeouts when there is high load
on the CI. This makes the integration of changes almost impossible.

As the previous solution to the timeouts stopped working a few weeks
ago and no new solution has been found, the freeopcua tests are disabled
for now.

Change-Id: I9d16fa55d8c6b45c017de10cee3facd643c76dd5
Task-number: QTBUG-66704
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>