Add batch read API to QOpcUaClient
The current API is focused on reading multiple attributes from the
same node. For every node, slots must be connected to signals.
This doesn't scale very well for a large number of nodes.

This patch adds an API to read attributes of different nodes in
one call. The user can select the node id, the attribute and an
index range for each value in the request.
The results are returned in a single signal emitted by
the client.

Change-Id: I9a716cd0317e32fe7fa8fd532e09696d7166ad47
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>