Return blob for all extension objects with binary or xml encoding
The previous fix didn't take extension objects with types not residing
in namespace 0 or with XML bodies into account.
This patch returns a QByteArray for all encoded extension objects that
can't be automatically decoded.

Returning QOpcUa::QExtensionObject values with type and encoding
information will be added on the dev branch.

Change-Id: I5f130e7cea6433444e0fbc38ec76cb67381cefc8
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>