Improve the open62541 value converter
The scalar conversion is changed to use direct values instead of
QVariant as input. This allows using the conversion in other
places of the code without having to construct or decode a QVariant.

Conversions from QVariant to UA_Variant print a warning and return
an empty UA_Variant if the type parameter does not match the content
of the QVariant.

Change-Id: I18e5b915c84c2a45bb8e29357cbe907aaf86a7ed
Reviewed-by: Rainer Keller <Rainer.Keller@qt.io>