Add expanded node id resolution and node() overload
This patch adds an easy way to create a QOpcUaNode from an expanded
node id.
A nullpointer is returned if the node id is not local to the server
or if the namespace URI can't be resolved.

The client now calls updateNamespaceArray() after every connect,
manual updates are only required if a change in the server's namespace
array is suspected.

Change-Id: Ia8aff90f247ea21f2bfb5369d7604a1175f2d068
Reviewed-by: Rainer Keller <Rainer.Keller@qt.io>