Use a counter for the node registration handle
Deleting a QOpcUaNodeImpl and creating a new one directly after has a
high probability for the new object to have the same address as the old
one.

This patch introduces a counter to provide handles for newly created
node objects. If registering a new node fails, QOpcUaClient::node()
returns a null pointer.

Change-Id: Iee6fb84e78e81e96e8f95c00c0428835e6db4523
Reviewed-by: Rainer Keller <Rainer.Keller@qt.io>