Add generated header and source files with known node ids
The NodeIds.csv file distributed by the OPC Foundation contains
the node ids of the default nodes in namespace 0.

This patch adds a generated header file with all node ids from
NodeIds.csv and a generated .cpp file with documentation.

The generator used to generate the files has been added in the
tools directory.

Change-Id: Iec8d58ca71fa5d29e691b064ec8bacf0abcd94d2
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>