Add memory management to the open62541 plugin
This patch adds a template class which calls a std::function
on a pointer in the destructor. This can be used to call
UA_<type>_delete() or UA_<type>_deleteMembers() on a value
when it goes out of scope.

A second template class deletes an array of open62541 types
in the destructor.

This makes the deallocation of values in functions with multiple
possible exit points much easier.

Change-Id: Ic47e094eb909bcf514ceb70c68904fa0a134642b
Reviewed-by: Rainer Keller <Rainer.Keller@qt.io>