Emit attributeUpdated whenever the attribute cache is updated
This allows to handle all reads, writes and data changes with
just a single slot.
The dataChangeOccurred signal is added to QOpcUaNode and is only
emitted after a data change notification has been received.

This is a change in behavior compared to 5.11 where the
attributeUpdated() signal was only emitted after a data change
and three signals were required to fully keep track of changes
to the attribute cache.

Change-Id: I3d7d33b7b0eb48b044d64178d448bedc4f3a756d
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>