Convert QLocalizedText to a gadget
This will not add any overhead to execution but saves a lot of QML hassle
because the wrapper that was present before needs to have additional
converter functions to allow QML to unwrap LocalizedText to QLocalizedText
in case it is returned in a QVariant. This happens for example if a node
method called on a server returns such types.

Change-Id: I664c417d1bebdc266764d51d564d12d98a4101a9
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>