configure: adjust to qtbase changes
the system now actually looks up libraries and headers (in as far as
declared, which is so far not the case here for headers), so any
additional paths need to be injected before forwarding to the inline
source's callback.

as a side effect, the handling of include paths is somewhat fixed -
previously, it would have added bogus paths if neither prefix nor
incdir was provided.

Fixes: QTBUG-72561
Change-Id: I2fb9c6f6b5053ea47a90e7439bbe4dca85fe1860
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>