Workaround for bug in Open62541
When cross-compiling on Winx64MSVC17 for Winx86MSVC17 the test server does
not start. There is a memcpy called with a source pointer of 0x00.

Happens in copy_noInit. The initial pointer is ok but due to the magic
pointer arithmetic the calcuation is afterwards pointing to a dereferenced
memory location that cointains a null pointer.
Actual issue happens when going into the else branch in that function.

Change-Id: I4dd78a114b1b1af88e995f4ce77aa9b87a35e4f5
Fixes: QTBUG-75020
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>