CMake: Fix CMake build for open62541
Fix -Wno flags to be assigned to the correct property as a list of
values.

Remove unnecessary Q_DECLARE_METATYPE because the type is already
registered automatically somewhere.

Fix standalone example using private API and not having the C language
enabled.

Remove optimization flags for open62541.c like the qmake project does,
to circumvent some miscompilation on MSVC which makes the test server
not start, and thus hangs CI integrations.
We do this for the test and example, but not the plugin, just to be
safe for now. It's done by a public OpcUA function called
qt_opcua_disable_optimizations_in_current_dir.

Regenerate some projects that used old APIs.

Task-number: QTBUG-85939
Change-Id: I2aea0e64371e99da3e2330f477c754728a8a5dd5
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Jannis Völker <jannis.voelker@basyskom.com>