Improve open62541 client iterate timer
UA_Client_run_iterate() is now called independently of existing
subscriptions. If there is no subscription, it will be called every
5000ms. Else, the interval is between the configured shortest interval
and 5000ms, depending on the shortest publishing interval.

The backend parameter "minimumClientIterateIntervalMs" can be used to
modify the minimum value.

The function is now called with 10ms of additional processing time, so
the keepalive mechanism can be enabled.

These changes reduce the CPU load caused by the open62541 backend.

Change-Id: I454a81fa7c4f4475a3cf73843e8781e3d3da4017
Pick-to: 5.14 5.15
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>