Add data type code generator
This change adds a code generator which generates data classes,
enums and encoder/decoder methods for OPC UA data type descriptions
in a .bsd file.

The generated data classes provide a getter and setter based API
that follows the structure of the types available in Qt OPC UA.
If a structure field's type is already available in Qt OPC UA,
it will be used in the data class instead of generating a new
data class for it.

The decoder/encoder operates on QOpcUaExtensionObject as well as
QByteArray and can be used to turn received binary encoded values
into easily accessible structured data as well as to create
binary encoded extension objects for writing custom types or for
parameters in method calls.

Change-Id: I8090dcb9b5f2d26d9ad02e3d6d5c45b3baa5c777
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>