Remove .qmake.conf from this repository
This file was necessary for the qmake build and the CI.  The qmake
build is history, and the CI does now read .cmake.conf (see
QTQAINFRA-4392).

In addition to being superfluous, the existence of .qmake.conf
triggers QTBUG-76140 when building Qt examples with qmake.  Removing
the file alleviates the symptoms of this bug.

Fixes: QTBUG-92271
Change-Id: Ib2b5ec2e7ebefd5453e6dc3d2b7913dff9c59a43
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>