Add support for the DataTypeDefinition attribute
This change extends the node attribute enum with the DataTypeDefinition
attribute and adds the necessary new types for reading and writing:

- QOpcUaStructureField
- QOpcUaStructureDefinition
- QOpcUaEnumField
- QOpcUaEnumDefinition

Reading and writing the new types is currently only supported in the open62541 plugin.
QOpcUaBinaryDataEncoding has been extended to encode and decode the new types.

The DataType attribute contains the full description of the data type and enabled
the user to generically decode binary encoded extension objects.

[ChangeLog][QOpcUaNode] The DataTypeDefinition attribute of DataType nodes can now
    be read. This allows manual decoding of generic structured and enum types.

Change-Id: I9ca5b3a8f30e7929e194b370d307261f6293747d
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>