Switch examples to build as isolated sub-builds part 2
qt_examples_build_begin needs the EXTERNAL_BUILD flag
to know that it's safe to build examples as ExternalProjects.

It still won't do it in CI until we enable building
examples as ExternalProjects for prefix builds.

This is preparation for that.

Task-number: QTBUG-90820
Change-Id: Iccf39b9cd7ba092a0ff050781905cf56068135fe
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 5416f3cbb3d037983273593dba9aad33795aa824)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>