Update commercial license headers
Updated header.COMM to the files in tqtc-qtopcua.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4971
Change-Id: Ia30ca705ab305627ecb86241d4955489c5d7e8e7
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>