Disable tests for Android for QtOpCua because they rely on QProcess
The tests are relying on QProcess to run a server that should be
packaged with the test, Android has a limitation that prevents
running such server with QProcess, thus the tests won't work.
Since it seems all tests are using that same mechanism here, that's
why disabling them seem better to avoid building, running and then
just skipping at the end.

Task-number: QTBUG-88846
Pick-to: 6.2 6.3
Change-Id: I50a08f03e2486f9fcf08f753e44a1641b650e326
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>