CMake: Fix No target "OpenSSL::SSL" error during prl generation
The OpenSSL non-global imported target would sometimes only be created
in the opcua module subdirectory scope, which made it unavailable for
prl processing in the plugin and qml directory scopes. Only sometimes,
because it depends on whether QtNetwork was built with openssl support
or not.

Move the find_package call higher in the directory scope tree so that
it can be seen by all libraries that need it.

In the case when QtNetwork is built with openssl support, the openssl
targets would be created in the root scope, which would cause target
global promotion issues.

To work around the issues, we first look for the openssl packages with
regular find_package, mark the targets not to be promoted, and then
use qt_find_package. The qt_find_package is needed to ensure proper
dependency registration in static builds, which was missing before due
to the missing PROVIDED_TARGETS option.

Pick-to: 6.4
Task-number: QTBUG-96283
Task-number: QTBUG-102883
Task-number: QTBUG-103905
Change-Id: I9f7249ebcf558007abfcd269b5f42b322e1fb17e
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>