MSVC: Avoid error about exceeding maximum number of constexpr steps
By default MSVC's step limit for constexpr is 100,000 steps (see
https://docs.microsoft.com/en-us/cpp/build/reference/constexpr-control-constexpr-evaluation?view=msvc-170).

With introduction of C++ constexpr for meta object string data
(7ceba9c4723846 in qtbase/dev) we clearly overstep this limit and run
into compiler errors while building.

Empirical research showed that a step limit of 2,500,000 is not enough
to fix the issue.

Fixes: QTBUG-105593
Change-Id: I783ec6301172840ff242f42253b9d7e4cc085fce
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>