OpcuaViewer: Don't locate files relative to binary
The logic to locate files relative to the binary is fragile. It is
currently broken for CMake on Windows, and not actually implemented
in qmake. Furthermore, it requires the binary directory to be
writable at runtime, because additional folders will be created
there.

Instead, embed the files via qrc, and extract them at runtime
into AppLocalDataLocation, which we know we can read/write to.

Fixes: QTBUG-109097
Change-Id: Ia6994339f4eba94ee6ef0cd82eadcc72563b7d47
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit ddfe3639386c235d50a54c46071983525ed3a645)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>