open62541: Re-encode unhandled types to QOpcUaExtensionObject
Depending on its configuration, the open62541 stack decodes several
data types we don't have Qt types for.
This patch adds the capability to re-encode them into an extension object
to allow the user to decode the data using QpcUaBinaryDataEncoding.

This makes the plugin more robust against different build variants
of open62541 and the current parameter set becomes a minimal requirement.

[ChangeLog][open62541 plugin] Unknown types that are decoded by the stack
    and have no Qt OPC UA data classes will now be re-encoded and returned
    as extension object instead of an empty QVariant.

Change-Id: I16408e6adf4f16e38430b32c6496c64b32becf62
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>