Handle OpenSSL with disabled SHA-1 signatures (open62541)
RHEL 9 ships the OpenSSL libraries with SHA-1 signatures disabled.
The security policies Basic128Rsa15 and Basic256 rely on SHA-1
which renders them unsupported on this platform.

This change checks for support and removes the two policies from
the list returned by QOpcUaClient::supportedSecurityPolicies()
and the endpoints in the test server if necessary.

Change-Id: I7bf9f903b159d794ef02163760d0c4c4781538e9
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>