QOpcUaConnectionSettings: use chrono types for timeouts
More type-safe, larger range, more intuitive literals (10min instead
of 10*60*1000).

Allows to use saturation arithmetic when handing the values to the
backend, instead of the incorrect truncation arithmetic that would
occur as part of the implicit conversion to quint32.

Found in API-review.

As a drive-by, fix a few grammar issues in the docs.

Pick-to: 6.6
Change-Id: Ie559202e6e48d1e07d0ee14fb7cbcc3e0c635a77
Reviewed-by: Jannis Völker <jannis.voelker@basyskom.com>
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>