Add the DiagnosticInfo data type
DiagnosticInfo is a built-in OPC UA type which is used together
with a string table in the response header to provide additional
diagnostics for a service call.

We don't expose the string table but we need this type to support
generic type decoding (for example, TransferResultErrorDataType
in the DI model contains a DiagnosticInfo field.

Reading and writing the new type is currently only supported in
the open62541 plugin. QOpcUaBinaryDataEncoding has been extended
to encode and decode the new types.

[ChangeLog][Qt OPC UA] The DiagnosticInfo data type is now available
    for reading and writing.

Change-Id: I3828d665e3450c0d6281d15da008d6c70b61c396
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>