Use std::optional instead of bool to indicate success of de/encoding
QOpcUaGenericStructHandler is a new class in Qt 6.7, and has two
functions that return values but might fail. The modern C++ way of
doing this is to return a std::optional, rather than return a bool with
the value returned through an out-parameter (or a default-
constructed value with a boolean out-parameter to indicate success).

Found during API review.

Pick-to: 6.7
Change-Id: Ib2ba6098f53eccab73beace6844028d0b674ca65
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>