Use QT_OPCUA_REMOVED_SINCE to get rid of the member operator==()
And replace them with hidden friends instead.

Rather than replacing with a hidden friend operator==(), that would
need to be exported as the class (hence its prior operator==() member)
is, export a hidden friend comparesEqual() and use this to implement the
old member operator.
This also simplifies the future migration to using comparison helper
macros.

As a drive-by: move the friend operators into the private section.

Found in 6.7 API review

Pick-to: 6.7
Change-Id: Idb0cc0f9b3a240f33e6a6c5844a09386c590276e
Reviewed-by: Marc Mutz <marc.mutz@qt.io>