CMake: Ensure the OpenSSL targets are global for SBOM generation
Before this change, if an OpenSSL target was not found in the root
scope (as a dependency of a static QtNetwork library), we would never
promote the OpenSSL target to global.

This causes issues for SBOM generation, where it expects system
library targets to be available in the scope of the project root
directory. In this case the target would be a non-global one confined
to the src subdirectory, not the root one.

Make sure we promote the OpenSSL target to global if it was not found
previously in the root scope, otherwise keep the existing logic not to
try double-promotion if the target was already found before.

Pick-to: 6.8
Task-number: QTBUG-122899
Change-Id: I3d525fe3f1e116878904cca9fcdfaa6d29cff6ae
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>