Fix race condition between backend shutdown and main thread
In rare cases, the backend shutdown attempts to log messages using
the test logger which leads to a heap-use-after-free.
This was caused by OpcUaConnection just calling deleteLater() on
the QOpcUaClient member in the destructor without waiting for it
to be destroyed.

As a solution, we delete the QOpcUaClient explicitly because its
destructor chain makes sure the backend thread is joined before
returning. The docs for the connection property of OpcUaConnection
state "Ownership of the client is transferred to QML", so we can
do that for all cases.

In addition, a potential connection error while the main thread
was waiting for the backend thread to join could lead to a deadlock
due to the attempt to emit the connectError() or
passwordForPrivateKeyRequired() signal connected to the main thread
via a BlockingQueuedConnection.

[ChangeLog][QML API] The connection property of OpcUaConnection now
    takes ownership of the QOpcUaClient as described in the docs

Task-number: QTBUG-126631
Pick-to: 6.6
Change-Id: I25de22ecbaa529af28d93244898395e68b065595
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>
(cherry picked from commit 247b3bae0732981869e2c45515fe24461e0a30b3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 8616972b2dc4409b72eaef8ae02053dc2abbbc02)