Fix anonymous authentication for secure connections
This change always adds the endpoint's security policy to the
authentication security policies, even if the selected token
type is anonymous.

Since open62541 1.4, this seems to be necessary or the connect
fails on the client side during activateSession.

As a drive-by, this change also adds the current opcuaviewer example
certificate to the trusted certificates of the open62541 server.

[ChangeLog][open62541 plugin] Fix a bug where anonymous
authentication over a secure connection was broken

Change-Id: I2bc2100a203857d8dfe3a294169ec4bbbc465a5e
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>
(cherry picked from commit 78f49db20e0102af9a99cf7b7980660d1fee229f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>