Doc: Remove examples.fileextensions line
Currently, the link to the simulationserver.h file is missing from the opc ua
example.

The .qdocconf contains the examples.fileextensions = ".cpp" line.
This line only includes example files that are of type .cpp.
This patch removes this line so that other file types, such as .h,
may be included as well.

Fixes: QTBUG-136141
Pick-to: 6.9 6.10
Change-Id: If2c13d29dc0341a23c5daad0919210c69197673b
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>