Doc: Fix incorrect usage of \generatelist
The \generatelist command takes an optional argument
that affect how the members are listed, for example,
`classesbymodule` which creates an annotated list of
the classes in a named C++ module.

`groupsbymodule` is not recognized or documented. QDoc
falls back to generating a list of all members of a
group in this case. While the end result is correct,
this relies on an undocumented feature that is likely
to change and break.

Replace these instances with an explicit \annotatedlist
command that produces the same output.

Task-number: QTBUG-138901
Change-Id: Id01a6f7401795cf2d318b8c74af1613bf3c28c81
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 700f29bc8a0b96b2ff99e95db5e32f9ae2a79c51)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>