Skip tests on macOS 26 in CI
For some reason, the tests cannot connect to the open62541-testserver
app, even though it is started. This needs to be investigated, but for
now let's simply skip these tests in order to enable macOS 26 in CI.

To skip C++ tests, this patch simply adds the relevant logic to each
initTestCase() call.

For the QML tests with custom setup classes it is a bit more
complicated - we need to do the check on the C++ side, and then
propagate the results to QML. Due to the fact that each QML test was
using the TestCase::when property, we needed to to update all the
conditions.

Task-number: QTBUG-139354
Change-Id: I1f1a223adff535a1dfeee7917cefa3cd9772568e
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>